+++ CHANGELOG.md
@@ -1,6 +1,39 @@
+<!--
+Copyright (C) 2025 Moko Consulting <hello@mokoconsulting.tech>
+This file is part of a Moko Consulting project.
+SPDX-License-Identifier: GPL-3.0-or-later
+
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation; either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see https://www.gnu.org/licenses/ .
+========================================================================
+FILE INFORMATION
+INGROUP: MokoDoliTools
+FILE: CHANGELOG.md
+VERSION: 02.05.05
+BRIEF: Changelog notes for the MokoDoliTools Dolibarr module
+PATH: /CHANGELOG.md
+========================================================================
+-->
+
 # CHANGELOG FOR MODULE MOKODOLITOOLS
 
-## 02.05.02
+## 02.05.05
+- Consolidated security access
+- Added missing language
+- Removed Adminer
+
+## 02.05.05
+- 
 - Rebrand from MokoCRM to MokoDoliTools
 - Removed "$this->conflictwith"